Refactor Duke to prepare for renaming
Place task, exception and command packages, as well as duke classes,
inside new duck package.
