+++ src/main/java/duck/command/AddCommand.java
@@ -1,10 +1,10 @@
-package command;
+package duke.command;
 
 import duke.TaskList;
-import task.Task;
+import duke.task.Task;
 
 /**
- * Represents the command that adds a task to the task list.
+ * Represents the duke.command that adds a task to the task list.
  */
 public class AddCommand extends Command {
     private final Task newTask;
@@ -20,7 +20,7 @@ public class AddCommand extends Command {
     }
 
     /**
-     * Executes the command of adding the task to the given task list.
+     * Executes the duke.command of adding the task to the given task list.
      *
      * @param taskList The task list to which the task is added.
      * @return String representing the task being added to the list.
