+++ src/main/java/duck/command/DeleteCommand.java
@@ -1,9 +1,9 @@
-package command;
+package duke.command;
 
 import duke.TaskList;
 
 /**
- * Represents the command that deletes a task from the task list.
+ * Represents the duke.command that deletes a task from the task list.
  */
 public class DeleteCommand extends Command {
     private final int index;
@@ -19,7 +19,7 @@ public class DeleteCommand extends Command {
     }
 
     /**
-     * Executes the command of deleting the task from the given task list.
+     * Executes the duke.command of deleting the task from the given task list.
      *
      * @param taskList The task list from which the task is deleted.
      * @return String representing the task being deleted.
