+++ src/main/java/duck/command/ShowScheduleCommand.java
@@ -1,11 +1,11 @@
-package command;
+package duke.command;
 
 import duke.TaskList;
 
 import java.time.LocalDate;
 
 /**
- * Represents the command that displays a schedule of tasks on a given date.
+ * Represents the duke.command that displays a schedule of tasks on a given date.
  */
 public class ShowScheduleCommand extends Command {
     private final LocalDate desiredDate;
@@ -21,7 +21,7 @@ public class ShowScheduleCommand extends Command {
     }
 
     /**
-     * Executes the command of showing the schedule of tasks on the given date.
+     * Executes the duke.command of showing the schedule of tasks on the given date.
      *
      * @param taskList The task list from which tasks taking place on the given date are taken to be listed.
      * @return String representing the list of tasks found.
