+++ src/main/java/duck/exception/DuckExceptionType.java
@@ -1,9 +1,9 @@
-package exception;
+package duck.exception;
 
 /**
  * An enumeration of all possible DukeExceptions.
  */
-public enum DukeExceptionType {
+public enum DuckExceptionType {
     CLOSE_INTERRUPT, DB_ADD, DB_DELETE, DB_DONE, DB_LAUNCH, DB_READ, INVALID_DATETIME, INVALID_FIND, INVALID_INPUT,
     INVALID_PERIOD, INVALID_SCHEDULE, INVALID_TASK_INDEX, MISSING_DEADLINE_DATETIME, MISSING_DEADLINE_DESC,
     MISSING_EVENT_DESC, MISSING_EVENT_PERIOD, MISSING_TODO_DESC
