+++ src/main/java/duck/task/Todo.java
@@ -1,4 +1,8 @@
-package task;
+package duke.task;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
 
 import java.time.LocalDate;
